requirements.txt
Flask-SQLAlchemy
SQLAlchemy
psycopg2
pyyaml
beanstalkc
slackweb
flask-httpauth
itsdangerous
